+++ seiscat/__init__.py
@@ -1,3 +1,14 @@
+# -*- coding: utf8 -*-
+# SPDX-License-Identifier: GPL-3.0-or-later
+"""
+Seiscat
+
+:copyright:
+    2022-2024 Claudio Satriano <satriano@ipgp.fr>
+:license:
+    GNU General Public License v3.0 or later
+    (https://www.gnu.org/licenses/gpl-3.0-standalone.html)
+"""
 from ._version import get_versions
 __version__ = get_versions()['version']
 del get_versions